separate pathway and status logic into individual modules
